put stake implementation in antoher file
